\name{exploreDDS}
\alias{exploreDDS}
\alias{exploreDDSplot}
\alias{hclustplot}
\alias{heatMaplot}
\alias{PCAplot}
\alias{MDSplot}
\alias{GLMplot}
\alias{MAplot}
\alias{tSNEplot}
\alias{volcanoplot}

\title{Visualization Methods for Differential Expression Analysis}

\description{These functions provide several visualization plots for differential expression analysis}

\usage{
exploreDDS(countMatrix, targets, cmp = cmp[[1]], preFilter = NULL, transformationMethod = "raw", blind = TRUE)
exploreDDSplot(countMatrix, targets, cmp = cmp[[1]],preFilter = NULL, samples, blind = TRUE,  scattermatrix = FALSE, plotly = FALSE, savePlot = FALSE, filePlot = NULL)
hclustplot(exploredds, method = "spearman", plotly = FALSE, savePlot = FALSE, filePlot = NULL)
heatMaplot(exploredds, clust, DEGlist = NULL, plotly = FALSE, savePlot = FALSE, filePlot = NULL, ...)
PCAplot(exploredds, plotly = FALSE, savePlot = FALSE, filePlot = NULL)
MDSplot(exploredds, method = "spearman", plotly = FALSE, savePlot = FALSE, filePlot = NULL)
GLMplot(exploredds, plotly = FALSE, savePlot = FALSE, filePlot = NULL, ...)
MAplot(degseqDF, FDR.cutoff = 0.05, comparison, filter = c(Fold = 2, FDR = 10), genes="NULL", plotly = FALSE, savePlot = FALSE, filePlot = NULL)
tSNEplot(countMatrix, targets, plotly = FALSE, savePlot = FALSE, filePlot = NULL, ...)
volcanoplot(degseqDF, comparison, filter = c(Fold = 2, FDR = 10), genes = "NULL", plotly = FALSE, savePlot = FALSE, filePlot = NULL)
}
\arguments{
\item{countMatrix}{\code{date.frame} or \code{matrix} containing raw read counts.}

\item{targets}{targets \code{data.frame}.}

\item{cmp}{\verb{character matrix} where comparisons are defined in two columns.
This matrix should be generated with the \code{\link[systemPipeR:readComp]{systemPipeR::readComp()}} 
function from the targets file. Values used for comparisons need to match those 
in the \code{Factor} column of the targets file.}

\item{preFilter}{allows removing rows in which there are very few reads. Accepts 
a numeric value with the minimum of total reads to keep. Default is \code{NULL}.}

\item{transformationMethod}{a \verb{character string} indicating which transformation 
method it will be used on the raw read counts. Supported methods include \code{rlog} 
and \code{vst} using the \code{DESeq2} package or default \code{raw} for no data 
transformation.}

\item{blind}{logical, whether to blind the transformation to the experimental 
design (see varianceStabilizingTransformation), from \code{\link[DESeq2:vst]{DESeq2::vst()}} 
or \code{\link[DESeq2:rlog]{DESeq2::rlog()}}.}

\item{samples}{a \code{character vector} of two samples or \code{ALL} samples in 
the dataset. Could be specified the \code{SampleName} column name of the targets 
file or the respective numeric values. Also, if set as \code{ALL}, a correlation
matrix it will be plot.}

\item{scattermatrix}{if \code{samples} set as \code{ALL}, requires to assign 
\code{TRUE} to build a correlation matrix and plot the correlogram of all the samples.}

\item{plotly}{logical: when \code{FALSE} (default), the \code{ggplot2} plot will
be returned. \code{TRUE} returns the \code{plotly} version of the plot.}

\item{savePlot}{logical: when \code{FALSE} (default), the plot will not be saved. 
If \code{TRUE} the plot will be saved, and requires the \code{filePlot} argument.}

\item{filePlot}{file name where the plot will be saved. For more information, 
please consult the \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} function.}

\item{exploredds}{object of class \code{\link[DESeq2:DESeqTransform]{DESeq2::DESeqTransform()}}.
\code{GLMplot} \code{exploredds} object of class \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}}, 
generated from \code{exploreDDS} function. Also, accepts the \code{date.frame} containing raw read counts.}

\item{method}{a \verb{character string} indicating which correlation coefficient
is to be computed, based on the \code{\link[stats:cor]{stats::cor()}} function. 
Options are: c("pearson" "kendall", "spearman").}

\item{clust}{select the data to apply the distance matrix computation. If \code{samples} 
selected, it will be applied the \code{\link[stats:dist]{stats::dist()}} function 
to the transformed count matrix to get sample-to-sample distances. If \code{ind}, 
it is necessary to provide the list of differentially expressed genes, for the 
\code{exploredds} subset.}

\item{DEGlist}{List of up or down regulated gene/transcript identifiers meeting
the chosen filter settings for all comparisons defined in data frames \code{pval} 
and \code{log2FC}.}

\item{FDR.cutoff}{filter cutoffs for the p-value adjusted.}

\item{degseqDF}{object of class \code{data.frame} generated by \code{\link[systemPipeR:run_edgeR]{systemPipeR::run_edgeR()}}
or \code{\link[systemPipeR:run_DESeq2]{systemPipeR::run_DESeq2()}}.}

\item{comparison}{\verb{character vector} specifying the factor names for comparison.}

\item{filter}{Named vector with filter cutoffs of format c(Fold=2, FDR=1) where
Fold refers to the fold change cutoff (unlogged) and FDR to the p-value cutoff.}

\item{genes}{\verb{character vecto}r of genes names to show on the plot.}

heatMaplot
\item{...}{For \code{heatMaplot} additional parameters for the \code{\link[pheatmap:pheatmap]{pheatmap::pheatmap()}} function.
For \code{GLMplot} additional parameters for the \code{\link[glmpca:glmpca]{glmpca::glmpca()}} function. 
For \code{tSNEplot} additional parameters for the \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} function.}
}

\details{
\code{exploreDDS} Convenience wrapper function to transform raw read counts using the \code{\link[DESeq2:DESeq2-package]{DESeq2::DESeq2-package()}} package transformations methods. The input file has to contain all the genes, not just differentially expressed ones.

\code{exploreDDSplot} Scatterplot of transformed counts from two samples or grid of all samples.

\code{hclustplot} This function computes the sample-wise correlation coefficients using the \code{\link[stats:cor]{stats::cor()}} function from the transformed expression values. After transformation to a distance matrix, hierarchical clustering is performed with the \code{\link[stats:hclust]{stats::hclust()}} function, and the result is plotted as a dendrogram.

\code{heatMaplot} This function performs hierarchical clustering on the transformed expression matrix. It uses, by default, a Pearson correlation-based distance measure and complete linkage for cluster join.

\code{PCAplot} This function plots a Principal Component Analysis (PCA) from transformed expression matrix. This plot shows samples variation based on the expression values and identifies batch effects.

\code{MDSplot} This function computes and plots multidimensional scaling analysis for dimension reduction of count expression matrix. Internally, it is applied the \code{\link[stats:dist]{stats::dist()}} function to the transformed count matrix to get sample-to-sample distances.

\code{GLMplot} This function computes and plots generalized principal components analysis for dimension reduction of count expression matrix.

\code{MAplot} This function plots log2 fold changes (y-axis) versus the mean of normalized counts (on the x-axis). Statistically significant features are colored.

\code{tSNEplot} This function computes and plots t-Distributed Stochastic Neighbor embedding (t-SNE) analysis for unsupervised nonlinear dimensionality reduction of count expression matrix. Internally, it is applied the \code{\link[Rtsne:Rtsne]{Rtsne::Rtsne()}} function, using the exact t-SNE computing with \code{theta=0.0}.

\code{volcanoplot} A simple function that shows statistical significance (\code{p-value}) versus magnitude of change (\verb{log2 fold change}).
}

\value{
\code{exploreDDS} function returns an object of class \code{\link[DESeq2:DESeqTransform]{DESeq2::DESeqTransform()}}.
\code{heatMaplot} function returns an object of \code{pheatmap} or \code{plotly} class.
\code{exploreDDSplot} function returns an object of \verb{ggplot2 plot}.
\code{hclustplot}, \code{PCAplot}, \code{MDSplot}, \code{GLMplot}, \code{MAplot}, 
\code{tSNEplot}, and \code{volcanoplot} functions returns an object of \code{ggplot} or \code{plotly} class.
}

\seealso{
Note that the recommendation is to use the resulting transformed values in the
\code{transformationMethod} argument only for visualization and clustering,
not for differential expression analysis which needs raw counts. Users are
strongly encouraged to consult the \code{\link[DESeq2:DESeq2-package]{DESeq2::DESeq2-package()}} vignette for more detailed
information on this topic and how to properly run \code{DESeq2} on data sets
with more complex experimental designs.
}

\examples{
## Targets file
targetspath <- system.file("extdata", "targets.txt", package="systemPipeR")
targets <- read.delim(targetspath, comment="#")
cmp <- systemPipeR::readComp(file=targetspath, format="matrix", delim="-")
## Count table file
countMatrixPath <- system.file("extdata", "countDFeByg.xls", package="systemPipeR")
countMatrix <- read.delim(countMatrixPath, row.names=1)

## Run exploredds
exploredds <- exploreDDS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="rlog")

## Plot exploreDDSplot
exploreDDSplot(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, samples=c(3,4))

## Plot hclustplot
hclustplot(exploredds, method = "spearman")

## Plot heatMaplot Samples
heatMaplot(exploredds, clust="samples", plotly = TRUE)

## Individuals genes identified in DEG analysis
### DEG analysis with `systemPipeR`
degseqDF <- systemPipeR::run_DESeq2(countDF = countMatrix, targets = targets, cmp = cmp[[1]], independent = FALSE)
DEG_list <- systemPipeR::filterDEGs(degDF = degseqDF, filter = c(Fold = 2, FDR = 10))
## Plot heatMaplot
heatMaplot(exploredds, clust="ind", DEGlist = unique(as.character(unlist(DEG_list[[1]]))))

## Plot PCAplot
PCAplot(exploredds, plotly = TRUE)

## Plot MDSplot
MDSplot(exploredds, plotly = FALSE)

## Plot GLMplot
exploredds_raw <- exploreDDS(countMatrix, targets, cmp=cmp[[1]], preFilter=NULL, transformationMethod="raw")
GLMplot(exploredds_raw, plotly = FALSE)

## Plot MAplot
MAplot(degseqDF, FDR.cutoff = 0.05, comparison = "M12-A12", filter = c(Fold = 1, FDR = 20), genes = "ATCG00280", plotly = TRUE)

## Plot tSNEplot
tSNEplot(countMatrix, targets, perplexity = 5)

## Plot volcanoplot
volcanoplot(degseqDF, comparison = "M12-A12", filter = c(Fold = 1, FDR = 20), genes = "ATCG00280")
}
\references{
For more details on \code{DESeq2}, please consult the following
page: \href{http://bioconductor.org/packages/release/bioc/html/DESeq2.html}{DESeq2}.
For more details on \code{targets} file definition, please consult the following
page: \href{http://www.bioconductor.org/packages/release/bioc/vignettes/systemPipeR/inst/doc/systemPipeR.html#25_structure_of_targets_file}{systemPipeR}.
}
\author{
Daniela Cassol
}
