\name{showDT}
\alias{showDT}
\title{Create an HTML table using DT package with fixed columns}
\description{
Create an HTML table using DT package with fixed columns
}
\usage{
showDT(data, ...)
}
\arguments{
\item{data}{data object (either a matrix or a data frame).}

\item{...}{Additional arguments used by datatable() function.}
}
\author{
Daniela Cassol
}
\examples{
showDT(iris)
}