\docType{package}
\name{systemPipeR-package}
\alias{systemPipeR-package}
\title{systemPipeR package for Workflow Environment}
\description{
The systemPipeR package provides a suite of R/Bioconductor for designing, building 
and running end-to-end analysis workflows on local machines, HPC clusters and cloud 
systems, while generating at the same time publication quality analysis reports.

For detailed information on usage, see the package vignette, by typing
vignette("systemPipeR"), or more information on the project here: https://systempipe.org/spr

All software-related questions should be posted to the Bioconductor Support Site: https://support.bioconductor.org

The code can be viewed at the GitHub repository: https://github.com/tgirke/systemPipeR
}
\references{
Backman TWH, Girke T (2016) systemPipeR: NGS workflow and report generation 
environment. BMC Bioinformatics 17 (1). https://doi.org/10.1186/s12859-016-1241-0
}
\author{
Daniela Cassol, Tyler Backman, Thomas Girke
}
\keyword{package}